UADS-62 refactor: Renamed components folder
