+++ web/src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Navbar from "../component/Navbar";
+import Navbar from "../components/Navbar";
 
 export default function Home() {
 	return (
